Refactor and remove code for flowchart page
