+++ Client/src/components/flowchart/flowchartSidePanel/courses/CourseSearchbar.tsx
