+++ Client/src/components/flowchart/flowchartSidePanel/courses/SidebarCourse.tsx
