+++ Client/src/components/flowchart/index.ts
@@ -8,14 +8,15 @@ export { default as FlowchartUnitCounts } from "./currentFlowchart/FlowchartUnit
 export { default as FlowchartOptions } from "./createFlowchart/FlowChartOptions";
 export { default as ProgressBar } from "./createFlowchart/ProgressBar";
 // Flowchart Log Components
-export { default as FlowchartLog } from "./flowchartSidebar/flowchartLog/FlowchartLog";
-export { default as FlowchartLogOptions } from "./flowchartSidebar/flowchartLog/FlowchartLogOptions";
+export { default as FlowchartLog } from "./flowchartSidePanel/flowchartList/FlowchartLog";
+export { default as FlowchartLogOptions } from "./flowchartSidePanel/flowchartList/FlowchartLogOptions";
 
-// Flowchart Sidebar Components
-export { SidebarFlowchart } from "./flowchartSidebar/SidebarFlowchart";
-export { default as CourseDropdown } from "./flowchartSidebar/courses/CourseDropdown";
-export { default as CourseSearchbar } from "./flowchartSidebar/courses/CourseSearchbar";
-export { default as SidebarCourse } from "./flowchartSidebar/courses/SidebarCourse";
+// Flowchart SidePanel Components
+export { default as CourseDropdown } from "./flowchartSidePanel/courses/CourseDropdown";
+export { default as CourseSearchbar } from "./flowchartSidePanel/courses/CourseSearchbar";
+export { default as SidebarCourse } from "./flowchartSidePanel/courses/SidebarCourse";
+export { default as FlowchartBuilderForm } from "./flowchartSidePanel/FlowchartBuilderForm";
+export { default as SavedFlowchartList } from "./flowchartSidePanel/flowchartList/SavedFlowchartList";
 // Utilities and Helpers
 export { default as defaultTermData } from "@/constants/flowPlaceholder";
 export {
