+++ Client/src/components/layout/FlowchartPage/MobileFlowchartLayout.tsx
@@ -1,4 +1,3 @@
-import { SidebarFlowchart } from "@/components/flowchart";
 import DragDropContextWrapper from "@/components/layout/dnd/DragDropContextWrapper";
 import FlowchartHeader from "@/components/layout/FlowchartPage/FlowchartHeader";
 import MobileHeader from "../dynamicLayouts/MobileHeader";
@@ -8,7 +7,6 @@ const MobileFlowchartLayout = ({ children }: { children: React.ReactNode }) => {
   const isNarrowScreen = useIsNarrowScreen();
   return (
     <DragDropContextWrapper>
-      <SidebarFlowchart />
       <div id="draggable-portal"></div>
       <div className="bg-slate-900 text-white min-h-screen flex flex-col overflow-hidden no-scroll w-full">
         {isNarrowScreen ? <MobileHeader /> : <FlowchartHeader />}
